Change make bridge commands extendable by using method attributes
Now you can extend the partial class outside the package further to add more commands of your own without the need to change anything inside the package
