+++ Assets/Plugins/WebGL/WebGLTools.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 039a8ffcf22dc8d4b8177667e1be714e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
