+++ Assets/Plugins/WebGL/WebGLTools/Supyrb.WebGLTools.asmdef
@@ -1,0 +1,17 @@
+{
+    "name": "Supyrb.WebGLTools",
+    "rootNamespace": "",
+    "references": [],
+    "includePlatforms": [
+        "Editor",
+        "WebGL"
+    ],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
