+++ Assets/Plugins/WebGL/WebGLTools/Supyrb.WebGLTools.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e6be6bd2cac63ce49a95472c9d6e313a
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
