+++ Assets/Plugins/WebGL/WebGLTools/WebGlBridge.Commands.cs
@@ -1,6 +1,6 @@
 ﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="WebGlBridge.cs">
-//   Copyright (c) 2021 Johannes Deml. All rights reserved.
+// <copyright file="WebGlBridge.Commands.cs">
+//   Copyright (c) 2022 Johannes Deml. All rights reserved.
 // </copyright>
 // <author>
 //   Johannes Deml
@@ -8,88 +8,13 @@
 // </author>
 // --------------------------------------------------------------------------------------------------------------------
 
-using System;
-using System.Reflection;
-using System.Text;
 using Supyrb.Attributes;
 using UnityEngine;
-using UnityEngine.Rendering;
 
 namespace Supyrb
 {
-	/// <summary>
-	/// Bridge to Unity to access unity logic through the browser console
-	/// </summary>
-	public class WebGlBridge : MonoBehaviour
+	public partial class WebGlBridge
 	{
-		[RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.BeforeSceneLoad)]
-		private static void OnBeforeSceneLoadRuntimeMethod()
-		{
-			SetGlobalVariables();
-			var bridgeInstance = new GameObject("WebGL");
-			DontDestroyOnLoad(bridgeInstance);
-			bridgeInstance.AddComponent<WebGlBridge>();
-		}
-
-		private static void SetGlobalVariables()
-		{
-			var graphicsDevice = SystemInfo.graphicsDeviceType;
-			string webGl = string.Empty;
-			switch (graphicsDevice)
-			{
-				case GraphicsDeviceType.OpenGLES2:
-					webGl = "WebGL 1";
-					break;
-				case GraphicsDeviceType.OpenGLES3:
-					webGl = "WebGL 2";
-					break;
-				default:
-					webGl = graphicsDevice.ToString();
-					break;
-			}
-			WebGlPlugins.SetVariable("webGlVersion", webGl);
-			WebGlPlugins.SetVariable("unityVersion", Application.unityVersion);
-		}
-
-		private void Start()
-		{
-			Debug.Log("Unity WebGL Bridge ready -> Run 'unityGame.SendMessage(\"WebGL\", \"Help\")' in the browser console to see usage");
-		}
-
-		[ContextMenu("Log all commands")]
-		[WebGlCommand(Description = "Log all available commands")]
-		public void Help()
-		{
-			StringBuilder sb = new StringBuilder();
-			MethodInfo[] methods = GetType().GetMethods(BindingFlags.Public | BindingFlags.Instance);
-
-			sb.AppendLine("Available unity interfaces:");
-			for (int i = 0; i < methods.Length; i++)
-			{
-				var method = methods[i];
-				WebGlCommandAttribute commandAttribute = method.GetCustomAttribute<WebGlCommandAttribute>();
-				if (commandAttribute != null)
-				{
-					sb.Append($"- {method.Name}(");
-					ParameterInfo[] parameters = method.GetParameters();
-					for (int j = 0; j < parameters.Length; j++)
-					{
-						var parameter = parameters[j];
-						sb.Append($"{parameter.ParameterType} {parameter.Name}");
-						if (j < parameters.Length - 1)
-						{
-							sb.Append(", ");
-						}
-					}
-
-					sb.AppendLine($") -> {commandAttribute.Description}");
-				}
-			}
-
-			sb.AppendLine("\nRun a command with 'unityGame.SendMessage(\"WebGL\", \"COMMAND_NAME\",PARAMETER)'");
-			Debug.Log(sb.ToString());
-		}
-
 		/// <summary>
 		/// Logs the current memory usage
 		/// Browser Usage: <code>unityGame.SendMessage("WebGL","LogMemory");</code>
@@ -148,11 +73,12 @@ namespace Supyrb
 		
 		/// <summary>
 		/// Toggle the visibility of the info panel in the top right corner
+		/// Browser Usage: <code>unityGame.SendMessage("WebGL", "ToggleInfoPanel");</code>
 		/// </summary>
 		[WebGlCommand(Description = "Toggle develop ui visibility of InfoPanel")]
 		public void ToggleInfoPanel()
 		{
 			WebGlPlugins.ToggleInfoPanel();
 		}
-	}
+	} 
 }
