+++ Assets/Plugins/WebGL/WebGLTools/WebGlBridge.Commands.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: a607faf85f83488d973cdb69730f4aa6
+timeCreated: 1663943852
