+++ Assets/Plugins/WebGL/WebGLTools/WebGlTimeTracker.cs
